backend/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
openai==1.0.0
azure-cognitiveservices-speech==1.28.0
azure-ai-translation-text==1.0.0
sounddevice==0.4.6
scipy==1.11.1
pydub==0.25.1
numpy==1.24.3
requests==2.31.0
